Add files via upload
I have completed first phase of the project. I have checked the accuracy and robustness values of the final version of dataset. It is pretty much robust in terms of R-Square (.87) and its accuracy values. Nonetheless, I will look at the ways and check the data preparation processes so as to increase the robustness in the second phase. It is actually an iterative process